'#1866: Use a base class in related items TableModels to simplify code.
