+++ iped-app/src/main/java/iped/app/ui/ReferencedByTableModel.java
@@ -18,16 +18,12 @@
  */
 package iped.app.ui;
 
-import java.awt.event.MouseEvent;
-import java.awt.event.MouseListener;
 import java.util.Arrays;
 import java.util.stream.Collectors;
 
 import javax.swing.ListSelectionModel;
 import javax.swing.SwingUtilities;
 import javax.swing.event.ListSelectionEvent;
-import javax.swing.event.ListSelectionListener;
-import javax.swing.table.AbstractTableModel;
 
 import org.apache.lucene.document.Document;
 
@@ -36,117 +32,15 @@ import iped.engine.search.IPEDSearcher;
 import iped.engine.search.LuceneSearchResult;
 import iped.engine.search.MultiSearchResult;
 import iped.engine.task.HashTask;
-import iped.engine.task.index.IndexItem;
 import iped.parsers.ares.AresParser;
 import iped.parsers.emule.KnownMetParser;
 import iped.parsers.shareaza.ShareazaLibraryDatParser;
 import iped.properties.BasicProps;
 import iped.properties.ExtraProperties;
 
-public class ReferencedByTableModel extends AbstractTableModel
-        implements MouseListener, ListSelectionListener, SearchResultTableModel {
+public class ReferencedByTableModel extends BaseTableModel {
 
-    /**
-     *
-     */
     private static final long serialVersionUID = 1L;
-
-    private LuceneSearchResult results = new LuceneSearchResult(0);
-    private int selectedIndex = -1;
-    private Document refDoc;
-
-    public void clear() {
-        results = new LuceneSearchResult(0);
-        fireTableDataChanged();
-    }
-
-    @Override
-    public int getColumnCount() {
-        return 3;
-    }
-
-    @Override
-    public int getRowCount() {
-        return results.getLength();
-    }
-
-    @Override
-    public String getColumnName(int col) {
-        if (col == 2)
-            return IndexItem.NAME;
-
-        return ""; //$NON-NLS-1$
-    }
-
-    @Override
-    public boolean isCellEditable(int row, int col) {
-        if (col == 1) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    @Override
-    public Class<?> getColumnClass(int c) {
-        if (c == 1) {
-            return Boolean.class;
-        } else {
-            return String.class;
-        }
-    }
-
-    @Override
-    public void setValueAt(Object value, int row, int col) {
-        App.get().appCase.getMultiBookmarks().setChecked((Boolean) value,
-                App.get().appCase.getItemId(results.getLuceneIds()[row]));
-        BookmarksController.get().updateUISelection();
-    }
-
-    @Override
-    public Object getValueAt(int row, int col) {
-        if (col == 0) {
-            return row + 1;
-
-        } else if (col == 1) {
-            return App.get().appCase.getMultiBookmarks()
-                    .isChecked(App.get().appCase.getItemId(results.getLuceneIds()[row]));
-
-        } else {
-            try {
-                Document doc = App.get().appCase.getSearcher().doc(results.getLuceneIds()[row]);
-                return doc.get(IndexItem.NAME);
-            } catch (Exception e) {
-                // e.printStackTrace();
-            }
-            return ""; //$NON-NLS-1$
-        }
-    }
-
-    @Override
-    public void mouseClicked(MouseEvent arg0) {
-    }
-
-    @Override
-    public void mouseEntered(MouseEvent arg0) {
-    }
-
-    @Override
-    public void mouseExited(MouseEvent arg0) {
-    }
-
-    @Override
-    public void mousePressed(MouseEvent arg0) {
-    }
-
-    @Override
-    public void mouseReleased(MouseEvent evt) {
-        if (evt.getClickCount() == 2 && selectedIndex != -1) {
-            int docId = results.getLuceneIds()[selectedIndex];
-            ExternalFileOpen.open(docId);
-        }
-
-    }
 
     @Override
     public void valueChanged(ListSelectionEvent evt) {
@@ -181,7 +75,6 @@ public class ReferencedByTableModel extends AbstractTableModel
 
         FileProcessor parsingTask = new FileProcessor(id, false);
         parsingTask.execute();
-
     }
 
     public void listReferencingItems(Document doc) {
@@ -192,32 +85,33 @@ public class ReferencedByTableModel extends AbstractTableModel
         String edonkey = doc.get(HashTask.HASH.EDONKEY.toString());
         String hashes = Arrays.asList(md5, sha1, sha256, edonkey).stream().filter(a -> a != null)
                 .collect(Collectors.joining(" "));
-        
+
         if (hashes.isEmpty()) {
             results = new LuceneSearchResult(0);
             refDoc = null;
         } else {
             String textQuery = ExtraProperties.LINKED_ITEMS + ":(" + hashes + ") ";
             textQuery += ExtraProperties.SHARED_HASHES + ":(" + hashes + ")";
-    
+
             try {
                 IPEDSearcher task = new IPEDSearcher(App.get().appCase, textQuery, BasicProps.NAME);
                 results = MultiSearchResult.get(task.multiSearch(), App.get().appCase);
-    
+
                 final int length = results.getLength();
-    
+
                 if (length > 0) {
                     SwingUtilities.invokeLater(new Runnable() {
                         @Override
                         public void run() {
-                            App.get().referencedByDock.setTitleText(Messages.getString("ReferencedByTab.Title") + " " + length);
+                            App.get().referencedByDock
+                                    .setTitleText(Messages.getString("ReferencedByTab.Title") + " " + length);
                         }
                     });
                     refDoc = doc;
                 } else {
                     refDoc = null;
                 }
-    
+
             } catch (Exception e) {
                 results = new LuceneSearchResult(0);
                 refDoc = null;
@@ -226,12 +120,5 @@ public class ReferencedByTableModel extends AbstractTableModel
         }
 
         fireTableDataChanged();
-
     }
-
-    @Override
-    public MultiSearchResult getSearchResult() {
-        return MultiSearchResult.get(App.get().appCase, results);
-    }
-
 }
