+++ iped-app/src/main/java/iped/app/ui/SubitemTableModel.java
@@ -19,14 +19,10 @@
 package iped.app.ui;
 
 import java.awt.Rectangle;
-import java.awt.event.MouseEvent;
-import java.awt.event.MouseListener;
 
 import javax.swing.ListSelectionModel;
 import javax.swing.SwingUtilities;
 import javax.swing.event.ListSelectionEvent;
-import javax.swing.event.ListSelectionListener;
-import javax.swing.table.AbstractTableModel;
 
 import org.apache.lucene.document.Document;
 
@@ -37,104 +33,9 @@ import iped.engine.task.index.IndexItem;
 import iped.properties.BasicProps;
 import iped.utils.LocalizedFormat;
 
-public class SubitemTableModel extends AbstractTableModel
-        implements MouseListener, ListSelectionListener, SearchResultTableModel {
+public class SubitemTableModel extends BaseTableModel {
 
-    /**
-     *
-     */
     private static final long serialVersionUID = 1L;
-
-    LuceneSearchResult results = new LuceneSearchResult(0);
-    int selectedIndex = -1;
-
-    @Override
-    public int getColumnCount() {
-        return 3;
-    }
-
-    @Override
-    public int getRowCount() {
-        return results.getLength();
-    }
-
-    @Override
-    public String getColumnName(int col) {
-        if (col == 2)
-            return IndexItem.NAME;
-
-        return ""; //$NON-NLS-1$
-    }
-
-    @Override
-    public boolean isCellEditable(int row, int col) {
-        if (col == 1) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    @Override
-    public Class<?> getColumnClass(int c) {
-        if (c == 1) {
-            return Boolean.class;
-        } else {
-            return String.class;
-        }
-    }
-
-    @Override
-    public void setValueAt(Object value, int row, int col) {
-        App.get().appCase.getMultiBookmarks().setChecked((Boolean) value,
-                App.get().appCase.getItemId(results.getLuceneIds()[row]));
-        BookmarksController.get().updateUISelection();
-    }
-
-    @Override
-    public Object getValueAt(int row, int col) {
-        if (col == 0) {
-            return row + 1;
-
-        } else if (col == 1) {
-            return App.get().appCase.getMultiBookmarks()
-                    .isChecked(App.get().appCase.getItemId(results.getLuceneIds()[row]));
-
-        } else {
-            try {
-                Document doc = App.get().appCase.getSearcher().doc(results.getLuceneIds()[row]);
-                return doc.get(IndexItem.NAME);
-            } catch (Exception e) {
-                // e.printStackTrace();
-            }
-            return ""; //$NON-NLS-1$
-        }
-    }
-
-    @Override
-    public void mouseClicked(MouseEvent arg0) {
-    }
-
-    @Override
-    public void mouseEntered(MouseEvent arg0) {
-    }
-
-    @Override
-    public void mouseExited(MouseEvent arg0) {
-    }
-
-    @Override
-    public void mousePressed(MouseEvent arg0) {
-    }
-
-    @Override
-    public void mouseReleased(MouseEvent evt) {
-        if (evt.getClickCount() == 2 && selectedIndex != -1) {
-            int docId = results.getLuceneIds()[selectedIndex];
-            ExternalFileOpen.open(docId);
-        }
-
-    }
 
     @Override
     public void valueChanged(ListSelectionEvent evt) {
@@ -158,10 +59,10 @@ public class SubitemTableModel extends AbstractTableModel
 
         String parentId = doc.get(IndexItem.ID);
 
-        String textQuery = IndexItem.PARENTID + ":" + parentId; //$NON-NLS-1$
+        String textQuery = IndexItem.PARENTID + ":" + parentId;
 
         String sourceUUID = doc.get(IndexItem.EVIDENCE_UUID);
-        textQuery += " && " + IndexItem.EVIDENCE_UUID + ":" + sourceUUID; //$NON-NLS-1$ //$NON-NLS-2$
+        textQuery += " && " + IndexItem.EVIDENCE_UUID + ":" + sourceUUID;
 
         try {
             IPEDSearcher task = new IPEDSearcher(App.get().appCase, textQuery, BasicProps.NAME);
@@ -174,7 +75,7 @@ public class SubitemTableModel extends AbstractTableModel
                     @Override
                     public void run() {
                         App.get().subitemDock.setTitleText(
-                                LocalizedFormat.format(sumSubitens) + Messages.getString("SubitemTableModel.Subitens")); //$NON-NLS-1$
+                                LocalizedFormat.format(sumSubitens) + Messages.getString("SubitemTableModel.Subitens"));
                     }
                 });
             }
@@ -185,12 +86,5 @@ public class SubitemTableModel extends AbstractTableModel
         }
 
         fireTableDataChanged();
-
     }
-
-    @Override
-    public MultiSearchResult getSearchResult() {
-        return MultiSearchResult.get(App.get().appCase, results);
-    }
-
 }
